Update to OpenAPI 3.0, upgrade Swagger UI, and switch to ReDOC
This updates the Swagger JSON generation script in package.json to
output the wiremock-admin-api.json in OpenAPI/Swagger 3.0 format, which
is supported in recent versions of api-spec-converter.

I also updated Swagger UI to the latest version, since the current
version doesn't support OpenAPI 3.0. Although I could've extracted the
new version in place, I figured it'd be better to handle it in
package.json by adding `swagger-ui-dist` as a dependency and copying it
over the necessary files during build time. This will make upgrades
easier in the future.